+++ CHANGELOG.md
@@ -1,7 +1,16 @@
-# 1.0.0
+# 1.0.2
 
-Initial version
+Added import/export methods to Hexastore objects:
+
+  - importString / exportString
+  - importJSON / exportJSON
+  - importZip / exportZip
+  - importNt / exportNt
 
 # 1.0.1
 
 Update Readme
+
+# 1.0.0
+
+Initial version